% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_calc_scans.R
\name{f_calc_scans}
\alias{f_calc_scans}
\title{Parabolic scan-count curve}
\source{
\url{https://github.com/Justin-Watkins/FOSBAAS/blob/master/R/f_calc_scans.R}
}
\usage{
f_calc_scans(x, y, j)
}
\arguments{
\item{x}{Location of the parabola's far root (controls its width); the curve
is scaled so that it has roots near \code{0} and near \code{x}.}

\item{y}{Peak height of the curve.}

\item{j}{The point (e.g. minute index) at which to evaluate the curve.}
}
\value{
A numeric scan count at point \code{j}.
}
\description{
Evaluates the parabola used to shape the gate \link{scan_data}: a curve that rises
to a peak and falls again, modelling the rush of fans arriving before a game.
}
\details{
Internally \verb{a = y / (x^2 - 300x + 300)} scales the parabola
\verb{a * (j^2 - 301j + 300)}.
}
\examples{
f_calc_scans(x = 230, y = 110, j = 150)

}
\seealso{
\code{\link[=f_get_scan_data]{f_get_scan_data()}}

Other operations: 
\code{\link{f_get_MMKN}()},
\code{\link{f_get_line_length}()},
\code{\link{f_get_scan_data}()},
\code{\link{f_get_scans}()},
\code{\link{f_get_third_degree_fit}()},
\code{\link{f_get_wait_times}()},
\code{\link{f_simulate_distribution}()}
}
\concept{operations}
