% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FOSBAAS.R
\docType{package}
\name{FOSBAAS}
\alias{FOSBAAS}
\title{FOSBAAS: A package to help with the book "Fundamentals of Sports Business and Strategy."}
\description{
The FOSBAAS package provides dozens of functions and data sets for practicing with sports data
}
\section{FOSBAAS functions}{

The FOSBAAS functions ...
}

