% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{freq_table_data_bin}
\alias{freq_table_data_bin}
\title{Binned frequency table of service times}
\format{
A data frame with 23 rows and 3 variables:
\describe{
\item{bin}{numeric; bin index.}
\item{mean_seconds}{numeric; mean service time (seconds) in the bin.}
\item{prob}{numeric; relative frequency of the bin (sums to 1).}
}
}
\source{
Simulated for the book. See \code{create_data/data_ch10_wait_times_data.R}.
}
\usage{
freq_table_data_bin
}
\description{
A binned empirical distribution of concession service times, used alongside
\link{freq_table_data} in the operations chapter.
}
\seealso{
Other datasets: 
\code{\link{aggregated_crm_data}},
\code{\link{customer_data}},
\code{\link{customer_renewals}},
\code{\link{demographic_data}},
\code{\link{fa_survey_data}},
\code{\link{freq_table_data}},
\code{\link{manifest_data}},
\code{\link{perceptual_data}},
\code{\link{scan_data}},
\code{\link{season_data}},
\code{\link{secondary_data}},
\code{\link{wait_times_data}},
\code{\link{wait_times_distribution_data}}
}
\concept{datasets}
\keyword{datasets}
