% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wait_times_distribution_data}
\alias{wait_times_distribution_data}
\title{Wait-time distribution sample}
\format{
A data frame with 3,000 rows and 5 variables:
\describe{
\item{transaction}{numeric; transaction index.}
\item{orderTimes}{numeric; ordering time, in seconds.}
\item{paymentTimes}{numeric; payment time, in seconds.}
\item{fulfillTimes}{numeric; fulfilment time, in seconds.}
\item{totalTime}{numeric; total time, in seconds.}
}
}
\source{
Simulated for the book. See \code{create_data/data_ch10_wait_times_data.R}.
}
\usage{
wait_times_distribution_data
}
\description{
A larger sample of concession wait times (same columns as \link{wait_times_data})
used to illustrate fitting and simulating from a distribution.
}
\seealso{
Other datasets: 
\code{\link{aggregated_crm_data}},
\code{\link{customer_data}},
\code{\link{customer_renewals}},
\code{\link{demographic_data}},
\code{\link{fa_survey_data}},
\code{\link{freq_table_data}},
\code{\link{freq_table_data_bin}},
\code{\link{manifest_data}},
\code{\link{perceptual_data}},
\code{\link{scan_data}},
\code{\link{season_data}},
\code{\link{secondary_data}},
\code{\link{wait_times_data}}
}
\concept{datasets}
\keyword{datasets}
