% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aggregated_crm_data}
\alias{aggregated_crm_data}
\title{aggregated_crm_data}
\format{
A data frame with 5000 rows and 3 variables:
\describe{
  \item{\code{repID}}{character rep responsible for calls}
  \item{\code{call}}{double number of calls}
  \item{\code{revenue}}{double revenue generated}
}
}
\usage{
aggregated_crm_data
}
\description{
Replicates aggregated call data
}
\details{
This data would be consolidated from raw CRM data
}
\keyword{datasets}
