% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{customer_data}
\alias{customer_data}
\title{customer_data}
\format{
A data frame with 200000 rows and 2 variables:
\describe{
  \item{\code{custID}}{character customer id}
  \item{\code{name}}{character customner name}
}
}
\usage{
customer_data
}
\description{
database of customer id and names
}
\details{
Replicates customer data found in a ticketing database
}
\keyword{datasets}
