% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{customer_renewals}
\alias{customer_renewals}
\title{customer_renewals}
\format{
A data frame with 13706 rows and 10 variables:
\describe{
  \item{\code{accountID}}{character customer account id}
  \item{\code{corporate}}{character is it a corporation}
  \item{\code{season}}{double season id}
  \item{\code{planType}}{character type of ticket purchased}
  \item{\code{ticketUsage}}{double percentage of tickets used}
  \item{\code{tenure}}{double years as a customer}
  \item{\code{spend}}{double amount spend to tickets}
  \item{\code{tickets}}{double number of tickets purchased}
  \item{\code{distance}}{double distance from arena}
  \item{\code{renewed}}{character did the customer renew their seats}
}
}
\usage{
customer_renewals
}
\description{
Simulates data for calculating renewal likelihood
}
\details{
This data would be transformed from many sources
}
\keyword{datasets}
