% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demographic_data}
\alias{demographic_data}
\title{demographic_data}
\format{
A data frame with 200000 rows and 13 variables:
\describe{
  \item{\code{custID}}{character Id from ticketing database}
  \item{\code{nameF}}{character First name}
  \item{\code{nameL}}{character Last name}
  \item{\code{nameFull}}{character Full name}
  \item{\code{gender}}{character gender}
  \item{\code{age}}{double age}
  \item{\code{latitude}}{double geographic latitude}
  \item{\code{longitude}}{double geographic longitude}
  \item{\code{distance}}{double distance from arena}
  \item{\code{maritalStatus}}{character married}
  \item{\code{ethnicity}}{character ethnicity}
  \item{\code{children}}{character children}
  \item{\code{hhIncome}}{double household income scaled}
}
}
\usage{
demographic_data
}
\description{
Replicates demographic information purchased on customers
}
\details{
Demographic data purchased from third parties takes this form
}
\keyword{datasets}
