% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_assign_promotion.R
\name{f_assign_promotion}
\alias{f_assign_promotion}
\title{f_assign_promotion}
\source{
\url{GIT}
}
\usage{
f_assign_promotion(number_bbhead, number_concert, number_other, games, seed)
}
\description{
Assign promotions to random dates
}
