% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_attend_coefficient.R
\name{f_attend_coefficient}
\alias{f_attend_coefficient}
\title{f_attend_coefficient}
\source{
\url{GIT}
}
\usage{
f_attend_coefficient(
  opponent,
  week_end,
  school_out,
  lag,
  opening_day,
  promotion,
  seed
)
}
\description{
Buikd a term to modify ticket sales
}
