% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_build_customer_ids_names.R
\name{f_build_customer_ids_names}
\alias{f_build_customer_ids_names}
\title{f_build_customer_ids_names}
\source{
\url{GIT}
}
\usage{
f_build_customer_ids_names(seed, num_rows = 2e+05)
}
\description{
Build a list of customers with names
}
