% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_build_demographic_data.R
\name{f_build_demographic_data}
\alias{f_build_demographic_data}
\title{f_build_demographic_data}
\source{
\url{GIT}
}
\usage{
f_build_demographic_data(
  seed1,
  seed2,
  mean_age_1,
  mean_age_2,
  sd_age_1,
  sd_age_2,
  seed3,
  seed4,
  seed5,
  seed6,
  seed7
)
}
\description{
combine demographic functions and build demographic data set
}
