% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_build_freq_table.R
\name{f_build_freq_table}
\alias{f_build_freq_table}
\title{f_build_freq_table}
\source{
\url{https://github.com/Justin-Watkins/FOSBAAS/blob/master/R/f_build_freq_table.R}
}
\usage{
f_build_freq_table(variable)
}
\arguments{
\item{variable}{A vector of factors}
}
\value{
creates a frequency table
}
\description{
Create a data frame of a frequency table
}
\examples{
f_build_freq_table()
}
