% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_build_season.R
\name{f_build_season}
\alias{f_build_season}
\title{f_build_season}
\source{
\url{GIT}
}
\usage{
f_build_season(
  seed1,
  season_year,
  seed2,
  num_games = 81,
  seed3,
  num_bbh,
  num_con,
  num_oth,
  seed4,
  seed5,
  mean_sales,
  sd_sales
)
}
\description{
Build season data
}
