% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_calculate_spend.R
\name{f_calculate_spend}
\alias{f_calculate_spend}
\title{f_calculate_spend}
\source{
\url{GIT}
}
\usage{
f_calculate_spend(corporate, planType, tenure, avgTenure)
}
\description{
Assign a spend per ticket amount
}
