% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_calculate_tenure.R
\name{f_calculate_tenure}
\alias{f_calculate_tenure}
\title{f_calculate_tenure}
\source{
\url{GIT}
}
\usage{
f_calculate_tenure(corporate, planType, distance, avgDist)
}
\description{
Assign a tenure amount in years
}
