% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_calculate_ticket_use.R
\name{f_calculate_ticket_use}
\alias{f_calculate_ticket_use}
\title{f_calculate_ticket_use}
\source{
\url{GIT}
}
\usage{
f_calculate_ticket_use(corporate, distance, avgDist)
}
\description{
What percentage of tickets get used
}
