% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_create_lead_scoring_data.R
\name{f_create_lead_scoring_data}
\alias{f_create_lead_scoring_data}
\title{f_create_lead_scoring_data}
\source{
\url{GIT}
}
\usage{
f_create_lead_scoring_data(
  seed,
  num_purchasers,
  season,
  f_tenure,
  f_spend,
  f_use,
  f_renewal_assignment,
  f_assign_renewal,
  renew = T
)
}
\arguments{
\item{seed}{PARAM_DESCRIPTION}

\item{num_purchasers}{How many purchasers do you want}

\item{season}{What year is the season you are describing}

\item{f_tenure}{function see ?f_calculate_tenure}

\item{f_spend}{function see ?f_calculate_spend}

\item{f_use}{function see ?f_calculate_ticket_use}

\item{f_renewal_assignment}{function see ?f_assign_renewal}

\item{f_assign_renewal}{function see ?f_assign_renewal}

\item{renew}{Do you want simulated renewals, Default: T}
}
\value{
A dataframe with 10 columns corresponding to features associated with season ticket holders
}
\description{
Build a dataset that resembles season ticket holder renewal data
}
