% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_days_since_last_game.R
\name{f_days_since_last_game}
\alias{f_days_since_last_game}
\title{f_days_since_last_game}
\source{
\url{GIT}
}
\usage{
f_days_since_last_game(dates)
}
\description{
Calculate the number of days since a game was played
}
