% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_demographics_children.R
\name{f_demographics_children}
\alias{f_demographics_children}
\title{f_demographics_children}
\source{
\url{GIT}
}
\usage{
f_demographics_children(seed, married, age)
}
\description{
Determine the ethnicity of the fan
}
