% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_demographics_ethnicity.R
\name{f_demographics_ethnicity}
\alias{f_demographics_ethnicity}
\title{f_demographics_ethnicity}
\source{
\url{GIT}
}
\usage{
f_demographics_ethnicity(seed, distance, age)
}
\description{
Determine the ethnicity of the fan
}
