% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_demographics_income.R
\name{f_demographics_income}
\alias{f_demographics_income}
\title{f_demographics_income}
\source{
\url{GIT}
}
\usage{
f_demographics_income(married, gender, age, age_low, age_high)
}
\description{
Determine the household income of the fan
}
