% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_demographics_married.R
\name{f_demographics_married}
\alias{f_demographics_married}
\title{f_demographics_married}
\source{
\url{GIT}
}
\usage{
f_demographics_married(seed, gender, age)
}
\description{
Determine if the customer is married
}
