% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_get_line_length.R
\name{f_get_line_length}
\alias{f_get_line_length}
\title{f_get_line_length}
\source{
\url{GIT}
}
\usage{
f_get_line_length(seed, n = 300, u1, sd1, u2, sd2)
}
\description{
Create a data frame to simulate line length
}
