% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_get_scan_data.R
\name{f_get_scan_data}
\alias{f_get_scan_data}
\title{f_get_scan_data}
\source{
\url{https://github.com/Justin-Watkins/FOSBAAS/blob/master/R/f_get_scan_data.R}
}
\usage{
f_get_scan_data(x_value, y_value, seed, sd_mod)
}
\arguments{
\item{seed}{random number}

\item{sd_mod}{standard deviation modifier}
}
\value{
Return a parabola based on different parameters
}
\description{
Return a data set that approximates scan data
}
\examples{
f_get_scan_data(1,5,714,20)
}
