% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_get_third_degree_fit.R
\name{f_get_third_degree_fit}
\alias{f_get_third_degree_fit}
\title{f_get_third_degree_fit}
\source{
\url{GIT}
}
\usage{
f_get_third_degree_fit(new_var, dist_fit)
}
\description{
Get third degree fit
}
