% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_get_third_degree_fit.R
\name{f_get_third_degree_fit}
\alias{f_get_third_degree_fit}
\title{f_get_third_degree_fit}
\source{
\url{https://github.com/Justin-Watkins/FOSBAAS/blob/master/R/f_get_third_degree_fit.R}
}
\usage{
f_get_third_degree_fit(new_var, dist_fit)
}
\value{
output of third degree polynomial function
}
\description{
Outputs a value based on a third degree polynomial function
}
\examples{
f_get_third_degree_fit()
}
