% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_get_time_observations.R
\name{f_get_time_observations}
\alias{f_get_time_observations}
\title{f_get_time_observations}
\source{
\url{https://github.com/Justin-Watkins/FOSBAAS/blob/master/R/f_get_time_observations.R}
}
\usage{
f_get_time_observations(start, end)
}
\arguments{
\item{start}{A number < 24}

\item{end}{A number < 24}
}
\value{
Formatted military time values
}
\description{
Helper function to format time
}
\examples{
f_get_time_observations(5,6)
}
