% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_get_wait_times.R
\name{f_get_wait_times}
\alias{f_get_wait_times}
\title{f_get_wait_times}
\source{
\url{GIT}
}
\usage{
f_get_wait_times(seed, n = 300, time, rate1, rate2, rate3)
}
\description{
Create a data frame to simulate concession waits
}
