% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_renewal_assignment.R
\name{f_renewal_assignment}
\alias{f_renewal_assignment}
\title{f_renewal_assignment}
\source{
\url{GIT}
}
\usage{
f_renewal_assignment(seed, sth_data, f_assign_renewal)
}
\description{
Determine whether the tickets will be renewed
}
