% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_select_team_for_series.R
\name{f_select_team_for_series}
\alias{f_select_team_for_series}
\title{f_select_team_for_series}
\source{
\url{GIT}
}
\usage{
f_select_team_for_series(seed, number)
}
\description{
Determine which team is played
}
