% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_simulate_distribution.R
\name{f_simulate_distribution}
\alias{f_simulate_distribution}
\title{f_simulate_distribution}
\source{
\url{https://github.com/Justin-Watkins/FOSBAAS/blob/master/R/f_simulate_distribution.R}
}
\usage{
f_simulate_distribution(observation, rand, seeds = 1)
}
\value{
Gets Wait time
}
\description{
Simulates wait time distribution
}
\examples{
f_simulate_distribution()
}
