% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_simulate_distribution.R
\name{f_simulate_distribution}
\alias{f_simulate_distribution}
\title{f_simulate_distribution}
\source{
\url{GIT}
}
\usage{
f_simulate_distribution(observation, rand, seeds = 1)
}
\description{
Simulate distribution
}
