% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fa_survey_data}
\alias{fa_survey_data}
\title{fa_survey_data}
\format{
A data frame with 10000 rows and 26 variables:
\describe{
  \item{\code{ReasonForAttending}}{character element rank}
  \item{\code{Socialize}}{double element rank}
  \item{\code{Tailgate}}{double element rank}
  \item{\code{TakeSelfies}}{double element rank}
  \item{\code{PostToSocialMedia}}{double element rank}
  \item{\code{SeeFriends}}{double element rank}
  \item{\code{VisitKidAttractions}}{double element rank}
  \item{\code{MeetMascot}}{double element rank}
  \item{\code{SnacksForKids}}{double element rank}
  \item{\code{KidsRunBases}}{double element rank}
  \item{\code{KidsSlide}}{double element rank}
  \item{\code{GetDinner}}{double element rank}
  \item{\code{EatParkFood}}{double element rank}
  \item{\code{SampleFood}}{double element rank}
  \item{\code{GetDrinks}}{double element rank}
  \item{\code{DrinkBeer}}{double element rank}
  \item{\code{BuyGear}}{double element rank}
  \item{\code{TourThePark}}{double element rank}
  \item{\code{VisitAttractions}}{double element rank}
  \item{\code{WatchPregameShow}}{double element rank}
  \item{\code{SeeTheChicken}}{double element rank}
  \item{\code{UpgradeSeats}}{double element rank}
  \item{\code{GetAutographs}}{double element rank}
  \item{\code{WatchGame}}{double element rank}
  \item{\code{SeePractice}}{double element rank}
  \item{\code{MeetPlayers}}{double element rank}
}
}
\usage{
fa_survey_data
}
\description{
Simulated survey response data to perform factor analysis
}
\details{
Fan asked to rank elements 0-10
}
\keyword{datasets}
