% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{manifest_data}
\alias{manifest_data}
\title{manifest_data}
\format{
A data frame with 44705 rows and 8 variables:
\describe{
  \item{\code{seatID}}{double unique id of seat}
  \item{\code{section}}{double section id}
  \item{\code{sectionNumber}}{double section number}
  \item{\code{rowNumber}}{double row number}
  \item{\code{seatNumber}}{double seat number}
  \item{\code{seasonPrice}}{double current season price}
  \item{\code{groupPrice}}{double current group price}
  \item{\code{singlePrice}}{double single game price}
}
}
\usage{
manifest_data
}
\description{
Data to simulate manifest data from a ticketing system
}
\details{
This data has been transformed and won't typically look like this
}
\keyword{datasets}
