% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{perceptual_data}
\alias{perceptual_data}
\title{perceptual_data}
\format{
A data frame with 3 rows and 10 variables:
\describe{
  \item{\code{Friendly}}{double count of repsonses}
  \item{\code{Exciting}}{double count of repsonses}
  \item{\code{Fresh}}{double double count of repsonses}
  \item{\code{Inovative}}{double double count of repsonses}
  \item{\code{Fun}}{double double count of repsonses}
  \item{\code{Old}}{double double count of repsonses}
  \item{\code{Historic}}{double double count of repsonses}
  \item{\code{Winners}}{double double count of repsonses}
  \item{\code{Great}}{double double count of repsonses}
  \item{\code{Expensive}}{double double count of repsonses}
}
}
\usage{
perceptual_data
}
\description{
Simulates Consolidated data useful for building perceptual maps
}
\details{
Consolidated data take from a survey
}
\keyword{datasets}
