% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scan_data}
\alias{scan_data}
\title{scan_data}
\format{
A data frame with 900 rows and 4 variables:
\describe{
  \item{\code{observations}}{double ordered observations}
  \item{\code{scans}}{double count of customer scans}
  \item{\code{action_time}}{double time of observation}
  \item{\code{date}}{character date of observation}
}
}
\usage{
scan_data
}
\description{
Replicates data periodically captured at and entry gate
}
\details{
This data replicates data captured by hand and a stopwatch at a gate
}
\keyword{datasets}
