% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{season_data}
\alias{season_data}
\title{season_data}
\format{
A data frame with 243 rows and 12 variables:
\describe{
  \item{\code{gameNumber}}{double game order}
  \item{\code{team}}{character away team}
  \item{\code{date}}{double game date}
  \item{\code{dayOfWeek}}{character day of the week}
  \item{\code{month}}{character month of event}
  \item{\code{weekEnd}}{logical was it a weekend}
  \item{\code{schoolInOut}}{logical were children in school}
  \item{\code{daysSinceLastGame}}{double days since last game was played}
  \item{\code{openingDay}}{logical is it opening day}
  \item{\code{promotion}}{character was there a promotion}
  \item{\code{ticketSales}}{double total tickets sold}
  \item{\code{season}}{double year of season}
}
}
\usage{
season_data
}
\description{
Simulates data aggregated for schedule comparisons
}
\details{
This data would be aggregated to perform regression analysis on schedules
}
\keyword{datasets}
