% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{secondary_data}
\alias{secondary_data}
\title{secondary_data}
\format{
A data frame with 853444 rows and 9 variables:
\describe{
  \item{\code{seatID}}{double manifest id of the seat}
  \item{\code{custID}}{character customer who purchased the tickets}
  \item{\code{ticketType}}{character single, season, group}
  \item{\code{gameID}}{double id of game purchased}
  \item{\code{tickets}}{double number of tickets purchased}
  \item{\code{priceKey}}{character seat and ticket type to match price}
  \item{\code{price}}{double original price of ticket}
  \item{\code{orderedCluster}}{double clustered ticket spend}
  \item{\code{secondayrPrice}}{double Price paid on secondary market}
}
}
\usage{
secondary_data
}
\description{
Simulates processed data from secondary market sales
}
\details{
It may be impossible to obtain this data
}
\keyword{datasets}
