% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wait_times_data}
\alias{wait_times_data}
\title{wait_times_data}
\format{
A data frame with 900 rows and 5 variables:
\describe{
  \item{\code{transaction}}{double Count of transactions}
  \item{\code{orderTimes}}{double order time in seconds}
  \item{\code{paymentTimes}}{double payment time in seconds}
  \item{\code{fulfillTimes}}{double fulfillment time in secongs}
  \item{\code{totalTime}}{double totla time in seconds}
}
}
\usage{
wait_times_data
}
\description{
Decomposed data captured at a concession stands
}
\details{
Replicates data gathered through direct observation
}
\keyword{datasets}
